feat: rename create endpoint to upsert-msr
